Separates image loader exceptions.
