+++ src/utils/exceptions.hpp
@@ -5,97 +5,6 @@
 #include <vector>
 
 namespace exceptions {
-#pragma region Image loader
-namespace loader {
-class InvalidTrainTestSplitException : public std::exception {
-  float split;
-
-  virtual const char *what() const throw() {
-    std::string s = "The train test split must be in the [0, 1], got " +
-                    std::to_string(split);
-    char *result = new char[s.length() + 1];
-    std::strcpy(result, s.c_str());
-    return result;
-  }
-
-public:
-  InvalidTrainTestSplitException(float split) : split(split){};
-};
-
-class InvalidDatasetException : public std::exception {
-  std::string dataset;
-
-  virtual const char *what() const throw() {
-    std::string s =
-        "An invalid dataset received. Expected \"train\" or \"test\", got " +
-        dataset;
-    char *result = new char[s.length() + 1];
-    std::strcpy(result, s.c_str());
-    return result;
-  }
-
-public:
-  InvalidDatasetException(std::string dataset) : dataset(dataset){};
-};
-
-class InvalidBatchSizeException : public std::exception {
-  int batchSize;
-
-  virtual const char *what() const throw() {
-    std::string s = "The batch size must be greater than or equal 1, got " +
-                    std::to_string(batchSize);
-    char *result = new char[s.length() + 1];
-    std::strcpy(result, s.c_str());
-    return result;
-  }
-
-public:
-  InvalidBatchSizeException(int batchSize) : batchSize(batchSize){};
-};
-
-class NoFilesFoundException : public std::exception {
-  std::string root, fileFormats;
-
-  virtual const char *what() const throw() {
-    std::string s = "No matching files were found at " + this->root +
-                    " with the extensions [" + this->fileFormats + "].";
-    char *result = new char[s.length() + 1];
-    std::strcpy(result, s.c_str());
-    return result;
-  }
-
-public:
-  NoFilesFoundException(const std::filesystem::path &root,
-                        const std::vector<std::string> &fileFormats)
-      : root(root) {
-    if (fileFormats.empty()) {
-      this->fileFormats = "";
-      return;
-    }
-    this->fileFormats = fileFormats[0];
-    for (int i = 1; i < fileFormats.size(); ++i) {
-      this->fileFormats += ", " + fileFormats[i];
-    }
-  };
-};
-
-class InvalidDataShapeAfterPreprocessingException : public std::exception {
-  int rows;
-
-  virtual const char *what() const throw() {
-    std::string s = "Data must have 1 row after preprocessing, got " +
-                    std::to_string(this->rows);
-    char *result = new char[s.length() + 1];
-    std::strcpy(result, s.c_str());
-    return result;
-  }
-
-public:
-  InvalidDataShapeAfterPreprocessingException(int rows) : rows(rows){};
-};
-} // namespace loader
-#pragma endregion Image loader
-
 #pragma region JSON
 namespace json {
 class JSONTypeException : public std::exception {
