+++ test/test_image_loader.cpp
@@ -1,6 +1,6 @@
+#include "exceptions/image_loader.hpp"
 #include "fixtures.hpp"
 #include "image_loader.hpp"
-#include "utils/exceptions.hpp"
 #include "utils/matrix.hpp"
 #include "utils/path.hpp"
 #include <gtest/gtest.h>
